Automatically set site_id to first available site.
This change automatically saves the first site_id to the UnifiClient
struct when new() is called. This reduces caller overhead, but limits us to the first site.
I'll need to add a feature to choose more than the first site if required. To implement this
I've moved logic from the main.rs over to the unifi.rs client constructor code.

Signed-off-by: Brendan Shephard <bshephar@bne-home.net>